updated the style again and reorganized the file structure
