+++ package.json
@@ -12,6 +12,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
+    "react-social-icons": "^5.4.1",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
