+++ src/App.css
@@ -1,4 +1,28 @@
-
 html  {
-  
+  background: linear-gradient(0.25turn, #3f87a6, #ebf8e1, #f69d3c);
+}
+
+
+.main-contain {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  flex-grow: 1;
+}
+
+
+.footer-contain  {
+  bottom: 0;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  opacity: 0.5;
+  padding: 10px;
+  position: absolute;
+  width:100%;
+}
+
+.footer-contain p:hover {
+  color: black;
+  cursor: pointer;
 }
